首页基本布局完毕

+++ static_antd/build.sh
@@ -1,7 +1,12 @@
 #!/usr/bin/env bash
 echo 'npm build successfully, start moving files.'
-mv dist/index.html ../app/view/index.html
-rm -rf ../app/public/antDesign
-mkdir ../app/public/antDesign
-mv dist/* ../app/public/antDesign/
-echo 'Done.'
+if [ -f "dist/index.html" ]
+then
+  mv dist/index.html ../app/view/index.html
+  rm -rf ../app/public/antDesign
+  mkdir ../app/public/antDesign
+  mv dist/* ../app/public/antDesign/
+  echo 'Done.'
+else
+  echo 'Not changed.'
+fi
